Implement Complex number operations in Java
Create a class Complex with the attribute real (int) and img (int). Create a default constructor that will initialize the attribute with the initial value 0.Create three instance of the class Complex and Implement some member function as below,

void input() : for taking the input of real and img.
void display(): for display the value of real and img.
void add(): to adding two complex numbers.
void sub(): to subtract two complex numbers.
void Multiplication(): to multiplying two complex numbers.